remove the deleted module from course, add security checks and rearrange the code

+++ Backend/src/controllers/module.controller.js
@@ -3,6 +3,7 @@ import { ApiResponse } from "../utils/ApiResponse.util.js";
 import { ApiError } from "../utils/ApiError.util.js";
 import { asyncHandler } from "../utils/asyncHandler.util.js";
 import { Course } from "../models/course.model.js";
+import mongoose from "mongoose";
 
 const createModule = asyncHandler(async (req, res) => {
     const { title, content, courseTitle } = req.body;
@@ -18,17 +19,17 @@ const createModule = asyncHandler(async (req, res) => {
     if (existedModule) {
         throw new ApiError(409, "Module with this title already exists");
     }
-
-    const module = await Module.create({
-        title,
-        content,
-    });
-
+    
     const course = await Course.findOne({ title: courseTitle });
 
     if(!course) {
         throw new ApiError(404, "Course not found");
     }
+
+    const module = await Module.create({
+        title,
+        content,
+    });
 
     course.modules.push(module._id);
     await course.save();
@@ -52,7 +53,16 @@ const getModuleById = asyncHandler(async (req, res) => {
 });
 
 const updateModule = asyncHandler(async (req, res) => {
+    if(!req?.user || !req.user?.isAdmin) {
+        throw new ApiError(403, "Unauthorized access");
+    }
+
     const { id } = req.params;
+
+    if(!id) {
+        throw new ApiError(400, "Missing ID in request")
+    }
+
     const { title, content } = req.body;
 
     if (
@@ -84,11 +94,27 @@ const updateModule = asyncHandler(async (req, res) => {
 const deleteModule = asyncHandler(async (req, res) => {
     const { id } = req.params;
 
+    if(!id) {
+        throw new ApiError(400, 'Missing Id in request')
+    }
+
+    const { courseTitle } = req.body;
+
+    const course = await Course.findOne({ title: courseTitle });
+
+    if(!course) {
+        throw new ApiError(500, "Server error while finding the related course");
+    }
+    
     const module = await Module.findByIdAndDelete(id);
     if (!module) {
         throw new ApiError(404, "Module not found");
     }
-
+    
+    // remove module id from course modules array
+    course.modules = course.modules.filter(itemId => itemId.toString() !== id);
+    await course.save();
+    
     return res
         .status(200)
         .json(new ApiResponse(200, "Module deleted successfully"));
